Initial commit.
form2json transforms the data collected by an HTML form into
a JSON object that serves as a base object to pass to
jQuery.ajax(); Optionally, it can return only the data.

Included jquery 1.3.2, as that's the current production version
I'm working with. The upgrade to 1.4.x should be inconsequential
to the plugin.