Make 'dataOnly' the default
This commit changes the default return value of the plugin.

Instead of returning the form data wrapped in an object ready to
be passed to jQuery.ajax, only a JSON representation of the form
data will be returned by default.

The wrapped version can be returned by supplying an arguments
object with the property 'wrapped' set to true.